% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wq_metadata.R
\name{create_wq_query}
\alias{create_wq_query}
\title{Create a Water Quality Query}
\usage{
create_wq_query(
  locations = "ALL",
  parameters = "ALL",
  methods = "ALL",
  projects = "ALL",
  matrices = "ALL",
  paramGroups = "ALL"
)
}
\arguments{
\item{locations}{Character vector of location names or "ALL"}

\item{parameters}{Character vector of parameter codes or "ALL"}

\item{methods}{Character vector of method codes or "ALL"}

\item{projects}{Character vector of project codes or "ALL"}

\item{matrices}{Character vector of matrix codes or "ALL"}

\item{paramGroups}{Character vector of parameter group codes or "ALL"}
}
\value{
A list containing the query structure
}
\description{
Create a Water Quality Query
}
\keyword{internal}
