% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get Locations Table}
\usage{
get_locations()
}
\value{
A data frame containing location information
}
\description{
Retrieves a table of locations from the API.
}
\examples{
locations <- get_locations()
head(locations)
}
