% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_metadata.R
\name{get_station_metadata}
\alias{get_station_metadata}
\title{Get Station Metadata}
\usage{
get_station_metadata(station_id)
}
\arguments{
\item{station_id}{A character string for the station ID}
}
\value{
A list containing metadata for the specified timeseries
}
\description{
Retrieves metadata for a specific station.
}
\examples{
get_station_metadata("LOX 10")
}
