% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_data.R
\name{get_timeseries_data}
\alias{get_timeseries_data}
\title{Get Timeseries Data}
\usage{
get_timeseries_data(dbkeys, startDate = NULL, endDate = NULL)
}
\arguments{
\item{dbkeys}{A character vector with the database keys to get timeseries data for}

\item{startDate}{Start date as string in "YYYY-MM-DD" format}

\item{endDate}{End date as string in "YYYY-MM-DD" format}
}
\value{
A tibble containing timeseries data
}
\description{
Retrieves timeseries data for specified database keys and date range.
}
\examples{
get_timeseries_data(dbkeys = "91510", startDate = "2020-01-01", endDate = "2024-01-01")
}
