% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wq_data.R
\name{get_wq_data}
\alias{get_wq_data}
\title{Get Water Quality Data}
\usage{
get_wq_data(..., timeseriesIds = NULL, startDate = NULL, endDate = NULL)
}
\arguments{
\item{...}{Query filters (locations, parameters, methods, projects, matrices, paramGroups) see \code{create_wq_query()}}

\item{timeseriesIds}{Character vector of timeseries IDs (optional)}

\item{startDate}{Start date as string in "YYYY-MM-DD" format}

\item{endDate}{End date as string in "YYYY-MM-DD" format}
}
\value{
A tibble of water quality values (NULL if no data found)
}
\description{
Retrieves water quality data based on specified filters and date range.
}
\examples{
get_wq_data(locations = "G722", parameters = "25", 
            startDate = "2020-01-01", endDate = "2024-01-01")
}
