% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wq_filters.R
\name{get_wq_filters}
\alias{get_wq_filters}
\title{Get Water Quality Filters}
\usage{
get_wq_filters()
}
\value{
A named list of filter options
}
\description{
Retrieves available filters for water quality data queries.
}
\examples{
filters <- get_wq_filters()
str(filters)
}
