% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wq_metadata.R
\name{get_wq_metadata}
\alias{get_wq_metadata}
\title{Get Water Quality Metadata}
\usage{
get_wq_metadata(
  ...,
  startDate = NULL,
  endDate = NULL,
  offset = 0,
  limit = 1000,
  sleep = 1
)
}
\arguments{
\item{...}{Query filters (locations, parameters, methods, projects, matrices, paramGroups) see \code{create_wq_query()}}

\item{startDate}{Start date as string in "YYYYMMDD" format}

\item{endDate}{End date as string in "YYYYMMDD" format}

\item{offset}{Integer specifying the number of records to skip}

\item{limit}{Integer specifying the maximum number of records to return per request}

\item{sleep}{Number of seconds to pause between paginated requests}
}
\value{
A tibble containing water quality metadata
}
\description{
Retrieves metadata for water quality measurements based on specified filters.
}
\examples{
get_wq_metadata(locations = c("G722", "BB52"), parameters = c("18", "25"),
                startDate = "20190101", endDate = "20191231")
}
